move Mark class
