+++ src/marks/delaunay.js
@@ -2,7 +2,7 @@ import {group, pathRound as path, select, Delaunay} from "d3";
 import {create} from "../context.js";
 import {Curve} from "../curve.js";
 import {constant, maybeTuple, maybeZ} from "../options.js";
-import {Mark} from "../plot.js";
+import {Mark} from "../mark.js";
 import {
   applyChannelStyles,
   applyDirectStyles,
