+++ src/marks/link.js
@@ -1,7 +1,7 @@
 import {pathRound as path} from "d3";
 import {create} from "../context.js";
 import {Curve} from "../curve.js";
-import {Mark} from "../plot.js";
+import {Mark} from "../mark.js";
 import {applyChannelStyles, applyDirectStyles, applyIndirectStyles, applyTransform} from "../style.js";
 import {markers, applyMarkers} from "./marker.js";
