+++ src/marks/text.js
@@ -16,7 +16,7 @@ import {
   isTextual,
   isIterable
 } from "../options.js";
-import {Mark} from "../plot.js";
+import {Mark} from "../mark.js";
 import {
   applyChannelStyles,
   applyDirectStyles,
