+++ src/marks/tick.js
@@ -1,6 +1,6 @@
 import {create} from "../context.js";
 import {identity, number} from "../options.js";
-import {Mark} from "../plot.js";
+import {Mark} from "../mark.js";
 import {applyDirectStyles, applyIndirectStyles, applyTransform, applyChannelStyles, offset} from "../style.js";
 
 const defaults = {
