+++ src/plot.js
@@ -1,17 +1,14 @@
 import {cross, group, sum, select, sort, InternMap, rollup} from "d3";
 import {Axes, autoAxisTicks, autoScaleLabels} from "./axes.js";
-import {Channel, Channels, channelDomain, valueObject} from "./channel.js";
+import {Channel} from "./channel.js";
 import {Context, create} from "./context.js";
-import {defined} from "./defined.js";
 import {Dimensions} from "./dimensions.js";
 import {Legends, exposeLegends} from "./legends.js";
-import {arrayify, isDomainSort, isScaleOptions, keyword, map, range, where, yes} from "./options.js";
-import {maybeNamed, maybeInterval} from "./options.js";
-import {maybeProject} from "./projection.js";
+import {Mark} from "./mark.js";
+import {arrayify, isScaleOptions, map, range, where, yes, maybeInterval} from "./options.js";
 import {Scales, ScaleFunctions, autoScaleRange, exposeScales} from "./scales.js";
 import {position, registry as scaleRegistry} from "./scales/index.js";
-import {applyInlineStyles, maybeClassName, maybeClip, styles} from "./style.js";
-import {basic, initializer} from "./transforms/basic.js";
+import {applyInlineStyles, maybeClassName} from "./style.js";
 import {consumeWarnings, warn} from "./warnings.js";
 
 /** @jsdoc plot */
@@ -22,7 +19,7 @@ export function plot(options = {}) {
   const className = maybeClassName(options.className);
 
   // Flatten any nested marks.
-  const marks = options.marks === undefined ? [] : options.marks.flat(Infinity).map(markify);
+  const marks = options.marks === undefined ? [] : flatMarks(options.marks);
 
   // Compute the top-level facet state. This has roughly the same structure as
   // mark-specific facet state, except there isn’t a facetsIndex, and there’s a
@@ -360,90 +357,34 @@ export function plot(options = {}) {
   return figure;
 }
 
-export class Mark {
-  constructor(data, channels = {}, options = {}, defaults) {
-    const {facet = "auto", fx, fy, sort, dx, dy, clip, channels: extraChannels} = options;
-    this.data = data;
-    this.sort = isDomainSort(sort) ? sort : null;
-    this.initializer = initializer(options).initializer;
-    this.transform = this.initializer ? options.transform : basic(options).transform;
-    if (facet === null || facet === false) {
-      this.facet = null;
-    } else {
-      this.facet = keyword(facet === true ? "include" : facet, "facet", ["auto", "include", "exclude"]);
-      this.fx = fx;
-      this.fy = fy;
-    }
-    channels = maybeNamed(channels);
-    if (extraChannels !== undefined) channels = {...maybeNamed(extraChannels), ...channels};
-    if (defaults !== undefined) channels = {...styles(this, options, defaults), ...channels};
-    this.channels = Object.fromEntries(
-      Object.entries(channels).filter(([name, {value, optional}]) => {
-        if (value != null) return true;
-        if (optional) return false;
-        throw new Error(`missing channel value: ${name}`);
-      })
-    );
-    this.dx = +dx || 0;
-    this.dy = +dy || 0;
-    this.clip = maybeClip(clip);
-  }
-  initialize(facets, facetChannels) {
-    let data = arrayify(this.data);
-    if (facets === undefined && data != null) facets = [range(data)];
-    if (this.transform != null) ({facets, data} = this.transform(data, facets)), (data = arrayify(data));
-    const channels = Channels(this.channels, data);
-    if (this.sort != null) channelDomain(channels, facetChannels, data, this.sort); // mutates facetChannels!
-    return {data, facets, channels};
-  }
-  filter(index, channels, values) {
-    for (const name in channels) {
-      const {filter = defined} = channels[name];
-      if (filter !== null) {
-        const value = values[name];
-        index = index.filter((i) => filter(value[i]));
-      }
-    }
-    return index;
-  }
-  // If there is a projection, and there are both x and y channels (or x1 and
-  // y1, or x2 and y2 channels), and those channels are associated with the x
-  // and y scale respectively (and not already in screen coordinates as with an
-  // initializer), then apply the projection, replacing the x and y values. Note
-  // that the x and y scales themselves don’t exist if there is a projection,
-  // but whether the channels are associated with scales still determines
-  // whether the projection should apply; think of the projection as a
-  // combination xy-scale.
-  project(channels, values, context) {
-    maybeProject("x", "y", channels, values, context);
-    maybeProject("x1", "y1", channels, values, context);
-    maybeProject("x2", "y2", channels, values, context);
-  }
-  scale(channels, scales, context) {
-    const values = valueObject(channels, scales);
-    if (context.projection) this.project(channels, values, context);
-    return values;
-  }
-  plot({marks = [], ...options} = {}) {
-    return plot({...options, marks: [...marks, this]});
-  }
+function plotThis({marks = [], ...options} = {}) {
+  return plot({...options, marks: [...marks, this]});
 }
+
+// Note: This side-effect avoids a circular dependency.
+Mark.prototype.plot = plotThis;
 
 /** @jsdoc marks */
 export function marks(...marks) {
-  marks.plot = Mark.prototype.plot;
+  marks.plot = plotThis;
   return marks;
 }
 
+function flatMarks(marks) {
+  return marks
+    .flat(Infinity)
+    .filter((mark) => mark != null)
+    .map(markify);
+}
+
 function markify(mark) {
-  return typeof mark?.render === "function" ? mark : new Render(mark);
+  return typeof mark.render === "function" ? mark : new Render(mark);
 }
 
 class Render extends Mark {
   constructor(render) {
-    super();
-    if (render == null) return;
     if (typeof render !== "function") throw new TypeError("invalid mark; missing render function");
+    super();
     this.render = render;
   }
   render() {}
